refactor: Refactoring Chat code
