+++ src/pages/ChatRoomPage.jsx
@@ -7,13 +7,11 @@ import {
   orderByChild,
   query,
   ref,
-  serverTimestamp,
   startAt,
   update,
 } from "firebase/database";
 import React, { useEffect, useRef, useState } from "react";
 import { useSelector } from "react-redux";
-import { useNavigate } from "react-router";
 import { useParams } from "react-router-dom";
 import ChatHeader from "../components/chat/ChatHeader";
 import ChatInput from "../components/chat/ChatInput";
@@ -21,105 +19,45 @@ import ChatMessage from "../components/chat/ChatMessage";
 import DateSeperationLine from "../components/chat/DateSeperationLine";
 import Spinner from "../components/common/atoms/Spinner";
 import "../firebase";
-import { convertToDate } from "../utils/convert";
 import useChatRoom from "../hooks/useChatRoom";
+import { convertToDate } from "../utils/convert";
 
 export default function ChatRoomPage() {
   const [messages, setMessages] = useState([]);
-  const [counterAvatar, setCounterAvatar] = useState(null);
-  const navigate = useNavigate();
   const [isLoading, setIsLoading] = useState(true); // [추가] 로딩 상태
   const { userInfo } = useSelector((state) => state.user);
   const { chatId } = useParams();
-  const [counterName, setCounterName] = useState("");
   const messageEndRef = useRef(null);
   let prevDate = null;
-  const { checkChatId } = useChatRoom();
+  const {
+    checkChatId,
+    checkChatRoomUser,
+    getCounterName,
+    updateReadMessage,
+    updateLastVisited,
+    counterAvatar,
+    counterName,
+  } = useChatRoom({ chatId, userId: userInfo.userId });
 
   useEffect(() => {
-    // 0. chatId 검증
-    checkChatId(chatId);
-
-    // 0. 채팅방 유저가 맞는지 검증
-    const checkChatRoomUser = async () => {
-      const chatRoomRef = ref(
-        getDatabase(),
-        `users/${userInfo.userId}/chatRooms/${chatId}`,
-      );
-      const chatRoom = await get(chatRoomRef);
-      if (!chatRoom.exists()) {
-        navigate("/404", { replace: true });
-        return false;
-      }
-      return true;
-    };
-
-    // 1. 저장되어 있는 메세지 가져오기
-    // 1-0) 상대방 아바타 가져오기
-    // 1-1) 채팅방 이름 가져오기
-    const getCounterName = async () => {
-      const snapShot = await get(
-        ref(getDatabase(), `users/${userInfo.userId}/chatRooms/${chatId}`),
-      );
-      setCounterName(snapShot.val().counterName);
-      const counterAvatarRef = ref(
-        getDatabase(),
-        `users/${snapShot.val().counterId}/avatar`,
-      );
-      const counterAvatarSnapShot = await get(counterAvatarRef);
-      setCounterAvatar(counterAvatarSnapShot.val());
-    };
-
-    // 1-2) 이전의 메세지 읽음 처리
-    const updateReadMessage = async () => {
-      const snapShot = await get(
-        ref(getDatabase(), `users/${userInfo.userId}/chatRooms/${chatId}`),
-      );
-      // 이전 메세지가 있을 경우
-      if (snapShot.exists()) {
-        const { lastVisited } = snapShot.val();
-        const q = query(
-          ref(getDatabase(), `messages/${chatId}`),
-          orderByChild("timestamp"),
-          startAt(lastVisited),
-        );
-        // 마지막 방문 시점 이후의 메세지
-        const snapShot2 = await get(q);
-        if (snapShot2.exists()) {
-          const updates = {};
-          Object.keys(snapShot2.val()).forEach((key) => {
-            updates[`/messages/${chatId}/${key}/isRead`] = true;
-          });
-          await update(ref(getDatabase()), updates);
-        }
-      }
-    };
-
-    // 1-3) 메세지 가져오기
+    checkChatId(); // 채팅방이 ID 검증
+    // 이전 메세지 가져오기
     const getMessages = async () => {
       const snapShot = await get(ref(getDatabase(), `messages/${chatId}`));
       setMessages(snapShot.val() ? Object.values(snapShot.val()) : []);
     };
 
-    // 1-4) 마지막 방문 시점 최신화(굳이 비동기를 기다릴 필요X)
-    const updateLastVisited = () => {
-      update(
-        ref(getDatabase(), `users/${userInfo.userId}/chatRooms/${chatId}`),
-        {
-          lastVisited: serverTimestamp(),
-        },
-      );
-    };
-
-    (async () => {
+    const startChat = async () => {
       if (await checkChatRoomUser()) {
-        await getCounterName();
-        await updateReadMessage();
-        await getMessages();
-        updateLastVisited();
+        await getCounterName(); // 상대방 정보 가져오기
+        await updateReadMessage(); // 읽음 처리
+        await getMessages(); // 읽음 처리 이후 메세지 가져오기
+        updateLastVisited(); // 마지막 방문 시점 최신화
         setIsLoading(false); // 메세지 가져오기 완료
       }
-    })();
+    };
+    // 1. 채팅방 시작
+    startChat();
 
     // 2. 메세지가 추가될 때마다 메세지 추가하기
     const sorted = query(
@@ -127,9 +65,9 @@ export default function ChatRoomPage() {
       orderByChild("timestamp"),
     );
     // 새로운 데이터가 추가될 때마다 변화를 감지하는 이벤트 리스너를 등록
-    // 1. 메세지 읽음 처리
-    // 2. 메세지 추가
-    // 3. 마지막 방문 시점 최신화
+    // 1). 메세지 읽음 처리
+    // 2). 메세지 추가
+    // 3). 마지막 방문 시점 최신화
     const unsubscribe = onChildAdded(
       query(sorted, startAt(Date.now())), // 현재 시간보다 큰 데이터만 가져오기
       (snapshot) => {
